services/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.4
h11==0.14.0
httptools==0.6.4
idna==3.10
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
pip==22.0.2
prometheus_client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.3.1
scipy==1.14.1
setuptools==59.6.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1